requirements.txt
strands-agents
bedrock-agentcore
bedrock-agentcore-starter-toolkit
boto3
python-dotenv
fastapi
uvicorn
